Implemented count to time conversion
created a function that will thake in the current minute count and convert the start time array to the proper value. This will account for variations in the  num of days per month and leap year. can be improved as the calculation is done each time. Next up is formatting the data string to be written to the SD Card