Implemented Pulse Alarm State!!!!
currently pulse alarm state is just printed to the debug window. LED state will be toggled in the future. Currently the LED is toggled while waiting for a BLE connection. In the future this will only blink for a short amount of time upon power cycle, then the alarm system can take over