Advertisement timout handler
Added a timer that generates a random time between two values.
If pending upload = true and advertisement timesout, we start this timer and re-start advertisements when it expires. random time so that if two devices turn on at the same time and timeout at the same time we dont get locked in a loop.