max re-tries and qspi sector fix.
Added a counter to limit the number of times the device witth re-try advertisements per upload.
added a method for incrementing the sector and moved it to main, it needs to be incremented when we collect enought data, not when we actually upload. If we dont connect we just keep writing ot the current sector for ever.